refactor pass 1
