+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-data-table-component",
-  "version": "0.4.3",
+  "version": "0.5.0",
   "description": "A declarative react based data table",
   "main": "dist/index.prod.js",
   "module": "dist/index.prod.js",
@@ -64,7 +64,7 @@
     "babel-eslint": "^8.2.2",
     "babel-jest": "^22.4.3",
     "babel-plugin-external-helpers": "^6.22.0",
-    "babel-plugin-module-resolver": "^3.1.0",
+    "babel-plugin-module-resolver": "^3.1.1",
     "babel-plugin-styled-components": "^1.5.1",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.26.0",
     "babel-preset-env": "^1.6.1",
@@ -99,12 +99,12 @@
     "rollup-plugin-node-resolve": "^3.3.0",
     "rollup-plugin-scss": "^0.4.0",
     "rollup-plugin-uglify": "^3.0.0",
+    "styled-components": "^3.2.3",
     "stylelint": "^9.1.3",
     "stylelint-config-standard": "^18.2.0",
     "stylelint-config-styled-components": "^0.1.1",
     "stylelint-processor-styled-components": "^1.3.1",
-    "uglify-es": "^3.3.9",
-    "styled-components": "^3.2.3"
+    "uglify-es": "^3.3.9"
   },
   "dependencies": {
     "lodash": "^4.17.5",
