+++ src/DataTable/TableRow.js
@@ -1,7 +1,10 @@
-import React from 'react';
+import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import styled, { withTheme, css } from 'styled-components';
 import TableCell from './TableCell';
+import TableCellCheckbox from './TableCellCheckbox';
+import TableCellExpander from './TableCellExpander';
+import { determineExpanderRowIdentifier } from './util';
 
 const TableRowStyle = styled.tr`
   border-top: 1px solid ${props => props.theme.rows.borderColor};
@@ -24,51 +27,91 @@ const TableRowStyle = styled.tr`
   `};
 `;
 
-const TableRow = ({ striped, highlightOnHover, pointerOnHover, children, columns, keyField, row, index, onRowClicked }) => {
-  const handleRowClick = e =>
-    onRowClicked && onRowClicked(row, index, e);
+class TableRow extends PureComponent {
+  static propTypes = {
+    striped: PropTypes.bool.isRequired,
+    highlightOnHover: PropTypes.bool.isRequired,
+    pointerOnHover: PropTypes.bool.isRequired,
+    columns: PropTypes.array.isRequired,
+    keyField: PropTypes.string.isRequired,
+    row: PropTypes.object.isRequired,
+    index: PropTypes.number.isRequired,
+    onRowClicked: PropTypes.func.isRequired,
+    onRowSelected: PropTypes.func.isRequired,
+    selectedRows: PropTypes.array.isRequired,
+    rows: PropTypes.array.isRequired,
+    checkboxComponent: PropTypes.oneOfType([
+      PropTypes.string,
+      PropTypes.node,
+      PropTypes.func,
+    ]).isRequired,
+    checkboxComponentOptions: PropTypes.object.isRequired,
+    selectableRows: PropTypes.bool.isRequired,
+    expandableRows: PropTypes.bool.isRequired,
+    onToggled: PropTypes.func.isRequired,
+  };
 
-  return (
-    <TableRowStyle
-      striped={striped}
-      highlightOnHover={highlightOnHover}
-      pointerOnHover={pointerOnHover}
-      onClick={handleRowClick}
-    >
-      {children}
-      {columns.map(col => (
-        <TableCell
-          type="cell"
-          key={`cell-${col.id}-${row[keyField] || index}`}
-          width={col.width}
-          column={col}
+  isExpanded = () => {
+    const rowIdentifier = determineExpanderRowIdentifier(this.props.row, this.props.keyField);
+    return this.props.rows.findIndex(r => rowIdentifier === r.parent) > -1;
+  }
+
+  handleRowChecked = r => this.props.onRowSelected && this.props.onRowSelected(r);
+
+  handleRowClick = e => this.props.onRowClicked && this.props.onRowClicked(this.props.row, this.props.index, e);
+
+  isChecked = () => this.props.selectedRows.indexOf(this.props.rows[this.props.index]) > -1;
+
+  render() {
+    const {
+      striped,
+      highlightOnHover,
+      pointerOnHover,
+      columns,
+      keyField,
+      row,
+      index,
+      checkboxComponent,
+      checkboxComponentOptions,
+      selectableRows,
+      expandableRows,
+      onToggled,
+    } = this.props;
+
+    return (
+      <TableRowStyle
+        striped={striped}
+        highlightOnHover={highlightOnHover}
+        pointerOnHover={pointerOnHover}
+        onClick={this.handleRowClick}
+      >
+        {selectableRows &&
+        <TableCellCheckbox
+          width="42px"
+          checked={this.isChecked()}
+          checkboxComponent={checkboxComponent}
+          checkboxComponentOptions={checkboxComponentOptions}
+          onClick={this.handleRowChecked}
           row={row}
-        />))}
-    </TableRowStyle>
-  );
-};
-
-TableRow.propTypes = {
-  striped: PropTypes.bool,
-  highlightOnHover: PropTypes.bool,
-  pointerOnHover: PropTypes.bool,
-  children: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.node),
-    PropTypes.node,
-  ]),
-  columns: PropTypes.array.isRequired,
-  keyField: PropTypes.string.isRequired,
-  row: PropTypes.object.isRequired,
-  index: PropTypes.number.isRequired,
-  onRowClicked: PropTypes.func,
-};
-
-TableRow.defaultProps = {
-  striped: false,
-  highlightOnHover: false,
-  pointerOnHover: false,
-  children: null,
-  onRowClicked: null,
-};
+        />}
+        {expandableRows &&
+        <TableCellExpander
+          onToggled={onToggled}
+          expanded={this.isExpanded()}
+          row={row}
+          index={index}
+        />}
+        {columns.map(col => (
+          <TableCell
+            type="cell"
+            key={`cell-${col.id}-${row[keyField] || index}`}
+            width={col.width}
+            column={col}
+            row={row}
+          />))}
+      </TableRowStyle>
+    );
+  }
+}
 
 export default withTheme(TableRow);
