+++ src/DataTable/__tests__/TableCell.test.js
@@ -6,7 +6,7 @@ import { shallowWithTheme } from '../../test-helpers';
 import TableCell from '../TableCell';
 
 test('component <TableCell /> should render correctly', () => {
-  const wrapper = shallowWithTheme(<TableCell />);
+  const wrapper = shallowWithTheme(<TableCell row={{}} column={{ selector: 'test' }} />);
 
   expect(wrapper.dive().dive()).toMatchSnapshot();
 });
